Changed default coordinate writing to not modify coordinates.
Previously, the default was to put the lowest xyz coordinate in the
bottom of the box. This is still the behaviour when using the command
line tool - but the option is now passed in through main.

Now, when printing intermediates for testing/figure making purposes,
the 0-centered coordinates are printed out by default

Refs #2